Improve error handling and fix *time_ms returning seconds

+++ src/models/secondary_device.rs
@@ -45,9 +45,9 @@ pub trait Device {
 }
 
 pub trait ArchiveWriter {
-    fn add_file(&mut self, file: &mut File, path: &PathBuf, ctime: u64, mtime: u64, size: u64);
-    fn add_directory(&mut self, path: &PathBuf, ctime: u64, mtime: u64);
-    fn add_symlink(&mut self, path: &PathBuf, ctime: u64, mtime: u64, target: &PathBuf);
+    fn add_file(&mut self, file: &mut File, path: &PathBuf, ctime: u128, mtime: u128, size: u64);
+    fn add_directory(&mut self, path: &PathBuf, ctime: u128, mtime: u128);
+    fn add_symlink(&mut self, path: &PathBuf, ctime: u128, mtime: u128, target: &PathBuf);
     fn finalize(&mut self, deleted_files: &Vec<PathBuf>, new_index: &Vec<u8>);
 }
